Add initial support for ISPC back-end to IvyScript
Current library caching is suboptimal as ISPC is not intended as a JIT compiler. This back-end depends on GCC for linking libraries.
Data transfer between host and device memory is currently not being performed, the back-end is functional with explicit synchronisation. See the saturation example.
As the processing files are now not limited to OpenCL execution, the file naming and structure will be adjusted in the future.